+++ airflow/post_install.sh
@@ -22,7 +22,6 @@ python $AIRFLOW_HOME/utils/set_default_variables.py \
     raw_table="raw_encounters_" \
     segment_info="segment_info" \
     spatial_measures_source="pipe_static.spatial_measures_20200311" \
-    distance_from_port_source="pipe_static.distance_from_port_20200311" \
     source_dataset="{{ var.value.PIPELINE_DATASET }}" \
     source_tables="position_messages_" \
     temp_bucket="{{ var.value.TEMP_BUCKET }}"